labelling_pipeline/llm_api/requirements.txt
fastapi
uvicorn
boto3
python-dotenv
psycopg2
pydantic
openai